ISPN-2089 Commit not working for distributed shared cache stores
Pretty much all DIST cache store operations should check the locality
of keys and take that into account when making decisions to operate
on the cache store, except one, clear. Treat this special case
differently in a explicit way to avoid breaking functionality in the
future.
