+++ core/src/main/java/org/infinispan/interceptors/CacheStoreInterceptor.java
@@ -23,7 +23,6 @@
 package org.infinispan.interceptors;
 
 import org.infinispan.commands.AbstractVisitor;
-import org.infinispan.commands.VisitableCommand;
 import org.infinispan.commands.tx.CommitCommand;
 import org.infinispan.commands.tx.PrepareCommand;
 import org.infinispan.commands.tx.RollbackCommand;
@@ -45,6 +44,7 @@ import org.infinispan.interceptors.base.JmxStatsCommandInterceptor;
 import org.infinispan.jmx.annotations.MBean;
 import org.infinispan.jmx.annotations.ManagedAttribute;
 import org.infinispan.jmx.annotations.ManagedOperation;
+import org.infinispan.loaders.CacheLoaderException;
 import org.infinispan.loaders.CacheLoaderManager;
 import org.infinispan.loaders.CacheStore;
 import org.infinispan.loaders.modifications.Clear;
@@ -116,7 +116,7 @@ public class CacheStoreInterceptor extends JmxStatsCommandInterceptor {
    /**
     * if this is a shared cache loader and the call is of remote origin, pass up the chain
     */
-   public final boolean skip(InvocationContext ctx, VisitableCommand command) {
+   protected boolean skip(InvocationContext ctx) {
       if (store == null) return true;  // could be because the cache loader does not implement cache store
       if ((!ctx.isOriginLocal() && loaderConfig.shared()) || ctx.hasFlag(SKIP_CACHE_STORE)) {
          log.trace("Skipping cache store since the cache loader is shared and we are not the originator.");
@@ -133,49 +133,50 @@ public class CacheStoreInterceptor extends JmxStatsCommandInterceptor {
 
    @Override
    public Object visitCommitCommand(TxInvocationContext ctx, CommitCommand command) throws Throwable {
-      if (!skip(ctx, command)) {
-         if (ctx.hasModifications()) {
-            // this is a commit call.
-            GlobalTransaction tx = ctx.getGlobalTransaction();
-            if (getLog().isTraceEnabled()) getLog().tracef("Calling loader.commit() for transaction %s", tx);
+      if (!skip(ctx))
+         commitCommand(ctx);
 
-            //hack for ISPN-586. This should be dropped once a proper fix for ISPN-604 is in place
-            Transaction xaTx = null;
-            if (transactionManager != null) {
-               xaTx = transactionManager.suspend();
-            }
-
-            try {
-               store.commit(tx);
-            } catch (Throwable t) {
-               throw t;
-            } finally {
-               // Regardless of outcome, remove from preparing txs
-               preparingTxs.remove(tx);
-
-               //part of the hack for ISPN-586
-               if (transactionManager != null && xaTx != null) {
-                  transactionManager.resume(xaTx);
-               }
-            }
-            if (getStatisticsEnabled()) {
-               Integer puts = txStores.get(tx);
-               if (puts != null) {
-                  cacheStores.getAndAdd(puts);
-               }
-               txStores.remove(tx);
-            }
-            return invokeNextInterceptor(ctx, command);
-         } else {
-            if (getLog().isTraceEnabled()) getLog().trace("Commit called with no modifications; ignoring.");
-         }
-      }
       return invokeNextInterceptor(ctx, command);
+   }
+
+   protected void commitCommand(TxInvocationContext ctx) throws Throwable {
+      if (ctx.hasModifications()) {
+         // this is a commit call.
+         GlobalTransaction tx = ctx.getGlobalTransaction();
+         if (getLog().isTraceEnabled()) getLog().tracef("Calling loader.commit() for transaction %s", tx);
+
+         //hack for ISPN-586. This should be dropped once a proper fix for ISPN-604 is in place
+         Transaction xaTx = null;
+         if (transactionManager != null) {
+            xaTx = transactionManager.suspend();
+         }
+
+         try {
+            store.commit(tx);
+         } finally {
+            // Regardless of outcome, remove from preparing txs
+            preparingTxs.remove(tx);
+
+            //part of the hack for ISPN-586
+            if (transactionManager != null && xaTx != null) {
+               transactionManager.resume(xaTx);
+            }
+         }
+         if (getStatisticsEnabled()) {
+            Integer puts = txStores.get(tx);
+            if (puts != null) {
+               cacheStores.getAndAdd(puts);
+            }
+            txStores.remove(tx);
+         }
+      } else {
+         if (getLog().isTraceEnabled()) getLog().trace("Commit called with no modifications; ignoring.");
+      }
    }
 
    @Override
    public Object visitRollbackCommand(TxInvocationContext ctx, RollbackCommand command) throws Throwable {
-      if (!skip(ctx, command)) {
+      if (!skip(ctx)) {
          if (getLog().isTraceEnabled()) getLog().trace("Transactional so don't put stuff in the cache store yet.");
          if (ctx.hasModifications()) {
             GlobalTransaction tx = ctx.getGlobalTransaction();
@@ -194,7 +195,7 @@ public class CacheStoreInterceptor extends JmxStatsCommandInterceptor {
 
    @Override
    public Object visitPrepareCommand(TxInvocationContext ctx, PrepareCommand command) throws Throwable {
-      if (!skip(ctx, command)) {
+      if (!skip(ctx)) {
          if (getLog().isTraceEnabled()) getLog().trace("Transactional so don't put stuff in the cache store yet.");
          prepareCacheLoader(ctx, command.getGlobalTransaction(), ctx, command.isOnePhaseCommit());
       }
@@ -204,7 +205,7 @@ public class CacheStoreInterceptor extends JmxStatsCommandInterceptor {
    @Override
    public Object visitRemoveCommand(InvocationContext ctx, RemoveCommand command) throws Throwable {
       Object retval = invokeNextInterceptor(ctx, command);
-      if (!skip(ctx, command) && !ctx.isInTxScope() && command.isSuccessful()) {
+      if (!skip(ctx) && !ctx.isInTxScope() && command.isSuccessful()) {
          Object key = command.getKey();
          boolean resp = store.remove(key);
          if (getLog().isTraceEnabled()) getLog().tracef("Removed entry under key %s and got response %s from CacheStore", key, resp);
@@ -214,17 +215,21 @@ public class CacheStoreInterceptor extends JmxStatsCommandInterceptor {
 
    @Override
    public Object visitClearCommand(InvocationContext ctx, ClearCommand command) throws Throwable {
-      if (!skip(ctx, command) && !ctx.isInTxScope()) {
-         store.clear();
-         if (getLog().isTraceEnabled()) getLog().trace("Cleared cache store");
-      }
+      if (!skip(ctx) && !ctx.isInTxScope())
+         clearCacheStore();
+
       return invokeNextInterceptor(ctx, command);
+   }
+
+   protected void clearCacheStore() throws CacheLoaderException {
+      store.clear();
+      if (getLog().isTraceEnabled()) getLog().trace("Cleared cache store");
    }
 
    @Override
    public Object visitPutKeyValueCommand(InvocationContext ctx, PutKeyValueCommand command) throws Throwable {
       Object returnValue = invokeNextInterceptor(ctx, command);
-      if (skip(ctx, command) || ctx.isInTxScope() || !command.isSuccessful()) return returnValue;
+      if (skip(ctx) || ctx.isInTxScope() || !command.isSuccessful()) return returnValue;
 
       Object key = command.getKey();
       InternalCacheEntry se = getStoredEntry(key, ctx);
@@ -238,7 +243,7 @@ public class CacheStoreInterceptor extends JmxStatsCommandInterceptor {
    @Override
    public Object visitReplaceCommand(InvocationContext ctx, ReplaceCommand command) throws Throwable {
       Object returnValue = invokeNextInterceptor(ctx, command);
-      if (skip(ctx, command) || ctx.isInTxScope() || !command.isSuccessful()) return returnValue;
+      if (skip(ctx) || ctx.isInTxScope() || !command.isSuccessful()) return returnValue;
 
       Object key = command.getKey();
       InternalCacheEntry se = getStoredEntry(key, ctx);
@@ -252,7 +257,7 @@ public class CacheStoreInterceptor extends JmxStatsCommandInterceptor {
    @Override
    public Object visitPutMapCommand(InvocationContext ctx, PutMapCommand command) throws Throwable {
       Object returnValue = invokeNextInterceptor(ctx, command);
-      if (skip(ctx, command) || ctx.isInTxScope()) return returnValue;
+      if (skip(ctx) || ctx.isInTxScope()) return returnValue;
 
       Map<Object, Object> map = command.getMap();
       for (Object key : map.keySet()) {
@@ -367,6 +372,7 @@ public class CacheStoreInterceptor extends JmxStatsCommandInterceptor {
 
    @ManagedAttribute(description = "number of cache loader stores")
    @Metric(displayName = "Number of cache stores", measurementType = MeasurementType.TRENDSUP)
+   @SuppressWarnings("unused")
    public long getCacheLoaderStores() {
       return cacheStores.get();
    }
